Add line at end of file
displayName is useful for debugging and required by [eslint-plugin-react](https://github.com/yannickcr/eslint-plugin-react) and [standard](https://github.com/feross/standard) which uses it. I figured this is something worth having. JSX [will infer it](https://facebook.github.io/react/docs/jsx-in-depth.html#the-transform) automatically from the variable assignment if undefined, but this allows you to notice and change it if needed.

Standard also requires a newline at the end of file.